Create class XMLParser to save and load melodies. Refactor MusicSheet…

+++ composeApp/src/commonMain/kotlin/org/goalteam/tunelint/model/core/impl/RestImpl.kt
@@ -3,7 +3,7 @@ package org.goalteam.tunelint.model.core.impl
 import org.goalteam.tunelint.model.core.Rest
 
 class RestImpl(
-    private val value: Float,
+    private val value: Int,
 ) : Rest {
     override fun value() = value
